fix(ingest): NA→float conversion + config_hash filter + JSON progress
bcolz_daily_bars.py: astype(float) before to_numpy() in to_ctable() so
pandas nullable-int NA values become NaN instead of pd.NA objects that
trigger "boolean value of NA is ambiguous" in winsorise_uint32.

duckdbdir.py:
- Strip config_hash/created_by identity cols from feature_cols before
  _compute_feature_metadata() so compute-internal columns don't get
  treated as data features for zipline.
- Emit structured JSON progress lines ({"type":"ingest_progress",...})
  on stdout from _pricing_iter() — per-batch for daily (batch_size=64),
  per-50-symbols for minute.  Enables Fyan orchestrator to forward
  per-symbol progress to web UI via SSE when running as subprocess.PIPE
  where click.progressbar is silent.

+++ src/zipline/data/bcolz_daily_bars.py
@@ -480,7 +480,9 @@ class BcolzDailyBarWriter:
         # Check all numeric columns (OHLC + numeric features + volume) together
         # Same as original Zipline: winsorise_uint32(raw_data, invalid_data_behavior, "volume", *OHLC)
         all_numeric_columns = all_scaled_columns + numeric_feature_columns_no_scale + ["volume"]
-        numeric_subset = raw_data[all_numeric_columns].to_numpy(copy=True)
+        # astype(float) converts pandas nullable-int NA → NaN before winsorise_uint32's
+        # `data[:, col_idx] > UINT32_MAX` comparison; avoids "boolean value of NA is ambiguous".
+        numeric_subset = raw_data[all_numeric_columns].astype(float).to_numpy(copy=True)
         
         # Apply negative offsets vectorized (faster than loop)
         if feature_offsets:
